Move some functionality of fileupload to FileUploader.tsx
Readdding FileUpload(er) In attempt to seperate concerns, and give ConfigSettingsLOQ a cleaner look)
